+++ test/libsolidity/smtCheckerTests/imports/duplicated_errors_1.sol
@@ -16,6 +16,7 @@ contract C is B {
 // SMTEngine: all
 // SMTSolvers: smtlib2
 // ----
+// Warning 2519: (c.sol:46-52): This declaration shadows an existing declaration.
 // Warning 3993: The BMC engine of the SMTChecker is deprecated and will be removed in a future release. Please use the CHC engine instead.
 // Warning 6328: (b.sol:62-75): CHC: Assertion violation might happen here.
 // Warning 3996: CHC analysis was not possible. No Horn solver was available. None of the installed solvers was enabled.